+++ calendar-service/src/main/java/com/ojosama/calendarservice/calendar/presentation/dto/request/CreateCalendarRequestDto.java
@@ -1,7 +1,5 @@
-package com.ojosama.calendarservice.calendar.presentaion.dto;
+package com.ojosama.calendarservice.calendar.presentation.dto.request;
 
-import com.fasterxml.jackson.annotation.JsonFormat;
-import com.fasterxml.jackson.annotation.JsonProperty;
 import com.ojosama.calendarservice.calendar.application.dto.command.CreateCalendarCommand;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
@@ -14,16 +12,14 @@ public record CreateCalendarRequestDto(
         @NotNull
         UUID eventId,
         @NotNull
-        @JsonProperty("startTime")
-        @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss.SSSSSS")
-        LocalDateTime eventStart
+        LocalDateTime eventDate
 ) {
     public CreateCalendarCommand toCommand(UUID userId) {
         return CreateCalendarCommand.builder()
+                .eventDate(this.eventDate)
                 .memo(this.memo)
                 .eventId(this.eventId)
                 .userId(userId)
-                .eventDate(this.eventStart)
                 .build();
     }
 }
